fix: update `Token` class to use `unsigned long`

+++ src/models/token.cpp
@@ -8,19 +8,19 @@
 // Реализация конструкторов
 Token::Token() : body(""), pos(0), index(0), fileId(0){}
 
-Token::Token(const std::string& body, long long pos, long long index,long long fileId)
+Token::Token(const std::string& body, unsigned long pos, unsigned long index,unsigned long fileId)
     : body(body), pos(pos), index(index), fileId(fileId){}
 
 // Геттеры
 std::string Token::getBody() const { return body; }
-long long Token::getPos() const { return pos; }
-long long Token::getIndex() const { return index; }
-long long Token::getFileId() const { return fileId; }
+unsigned long Token::getPos() const { return pos; }
+unsigned long Token::getIndex() const { return index; }
+unsigned long Token::getFileId() const { return fileId; }
 
 // Сеттеры
 void Token::setBody(const std::string& newBody) { body = newBody; }
-void Token::setPos(long long newPos) { pos = newPos; }
-void Token::setIndex(long long newIndex) { index = newIndex; }
+void Token::setPos(unsigned long newPos) { pos = newPos; }
+void Token::setIndex(unsigned long newIndex) { index = newIndex; }
 
 // Перегрузка оператора вывода в поток
 std::ostream& operator<<(std::ostream& os, const Token& token) {
