+++ src/models/token.h
@@ -7,25 +7,25 @@
 class Token {
 private:
     std::string body;  // Слово
-    long long pos;     // Позиция в тексте
-    long long index;   // Позиция среди токенов
-    long long fileId; //
+    unsigned long pos;     // Позиция в тексте
+    unsigned long index;   // Позиция среди токенов
+    unsigned long fileId; //
 
 public:
     // Конструкторы
     Token();  // Конструктор по умолчанию
-    Token(const std::string& body, long long pos, long long index, long long fileId);  // Параметризованный конструктор
+    Token(const std::string& body, unsigned long pos, unsigned long index, unsigned long fileId);  // Параметризованный конструктор
 
     // Геттеры
     std::string getBody() const;
-    long long getPos() const;
-    long long getIndex() const;
-    long long getFileId() const;
+    unsigned long getPos() const;
+    unsigned long getIndex() const;
+    unsigned long getFileId() const;
 
     // Сеттеры
     void setBody(const std::string& newBody);
-    void setPos(long long newPos);
-    void setIndex(long long newIndex);
+    void setPos(unsigned long newPos);
+    void setIndex(unsigned long newIndex);
 
     // Перегрузка оператора сравнения
     bool operator==(const Token& other) const;
